appengine/standard_python3/btc_price/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
google-cloud-secret-manager
google-cloud-bigquery
coinbase==2.1.0